SYNC-1569 Regenerate

+++ modules/apps/sync/sync-api/src/main/java/com/liferay/sync/service/SyncDeviceLocalServiceWrapper.java
@@ -85,7 +85,7 @@ public class SyncDeviceLocalServiceWrapper implements SyncDeviceLocalService,
 
 	@Override
 	public com.liferay.sync.model.SyncDevice addSyncDevice(long userId,
-		java.lang.String type, int buildNumber, int featureSet)
+		java.lang.String type, long buildNumber, int featureSet)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _syncDeviceLocalService.addSyncDevice(userId, type, buildNumber,
 			featureSet);
@@ -189,7 +189,7 @@ public class SyncDeviceLocalServiceWrapper implements SyncDeviceLocalService,
 
 	@Override
 	public com.liferay.sync.model.SyncDevice updateSyncDevice(
-		long syncDeviceId, java.lang.String type, int buildNumber,
+		long syncDeviceId, java.lang.String type, long buildNumber,
 		int featureSet, int status)
 		throws com.liferay.portal.kernel.exception.PortalException {
 		return _syncDeviceLocalService.updateSyncDevice(syncDeviceId, type,
